Rename and Buidl of GameManager to PointManager Class
Made the PointManager Class that tracks when a damageable object is destroyed (i.e. hp = 0). A DamageableObjects sends a OnObjectDeath event to the PointManager. The listener in the PointManager takes the MaxHP value from the argument e (The DamageableObjects object that died) and adds its MaxHP value to the currentScore variable in the Pointmanager and then Fires an event to the PointUIScript. The PointUIScript updates the UI to match the currentScore variable in the PointManager