Update to Point Manager Creation of Game Manager
Point Manager is now a class with a singleton pattern (only ever 1 instance in a scene). It tracks the number of active balls in a scene. When all balls are removed the point manager brings up the end of round screen.

The Ball class now adds itself to the list of active ball objects, allBalls, in the PointManager Script. The Ball class has an event that triggers when the ball dies in the FixedUpdate() function. The PointManager Listens to this event

The GameManager class handles Scene transitions and saving point values to PlayerPrefs. This has not yet been implemeneted.